Rewrite CA bundle updating script
The new script doesn't try to commit the changes. Different use cases
may want to handle committing differently, and it can be easily be done
externally.

The new script also updates the Changes file.

Update the github action to use the new script, and change to the
peter-evans/create-pull-request action, which is more commonly used.